Bugfix to navier-stokes transient setup files, change of default minimumminw in kernel_module.f90 (from 0.5 to 1.0) and addition of pruning script in misc to remove arb output and object files within all subdirectories.
Note that the change in minimumminw will increase the size of kernels being used.  In older (maybe v0.4ish) versions of the code minw = 1.d0 or larger.  When the default kernel polynomial order was changed from 1 to 2 (in v0.50), minw was decreased in an effort to keep the kernel sizes small (for the same minw, a polyorder=2 kernel will require more cells than a polyorder=1 kernel).  Testing on some unstructured mesh transient problems however has shown that using polyorder=2 + minw=0.5 can sometimes lead to instabilities, so minw has been increased again to 1.d0.  Note that the minw used can be set in the arb input file using 'KERNEL_OPTIONS minimumminw=0.5d0' for example.