Add warning for plot_alternate.py
plot_alternate.py is a workaround to work with newer versions of
matplotlib. This check exits the script if an older matplotlib version
is present. In the future, a single plot.py script will be able to deal
with all use cases -- the prototype for this is currently plot_merged.py